+++ ZoomChartsFacetChartPaid/release-notes.md
@@ -2,6 +2,18 @@
 
 =========================== Changelog ===========================
 
+Version 1.2.0.0 (2019-03-06)
+
+* Issue Number: BS-128.
+  Correct value formatting for primary and secondary axis.
+  Added new capabilities for primary and secondary axis:
+    - BETA - Axis Size
+    - BETA - Value Decimals
+    - BETA - Percentage Decimals
+* Added correct versioning start point.
+
+-----------------------------------------------------------------
+
 Version 1.1.1.2 (2019-03-05)
 
 * Issue Number: BS-143.